move this test to BuildersTest as the logic determining if a builder should run has been moved to Builders
